Refactor sorted multi
